updated the validation and the error messages and the controllers return results

+++ HomeCompassApi/Controllers/Cases/HomelessController.cs
@@ -19,27 +19,30 @@ namespace HomeCompassApi.Controllers.Cases
         [HttpGet]
         public ActionResult<List<Homeless>> Get()
         {
-            return _repository.GetAll().ToList();
+            return Ok(_repository.GetAll().ToList());
         }
 
         [HttpGet("{id}")]
         public ActionResult<Homeless> Get(int id)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var homeless = _repository.GetById(id);
             if (homeless is null)
-                return NotFound();
+                return NotFound($"There is no record with the specified Id: {id}");
 
-            return homeless;
+            return Ok(homeless);
         }
 
         [HttpPost]
         public IActionResult Create(Homeless homeless)
         {
-            if (homeless is null)
-                return BadRequest("Please provide the Homeless...");
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
 
             if (_repository.GetById(homeless.Id) is null)
-                return NotFound($"There is no record with the specified Id {homeless.Id}");
+                return NotFound($"There is no record with the specified Id: {homeless.Id}");
 
             _repository.Add(homeless);
             return CreatedAtAction(nameof(Get), new { id = homeless.Id }, homeless);
@@ -48,11 +51,11 @@ namespace HomeCompassApi.Controllers.Cases
         [HttpPut]
         public IActionResult Update(Homeless homeless)
         {
-            if (homeless is null)
-                return BadRequest();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
 
             if (_repository.GetById(homeless.Id) is null)
-                return NotFound();
+                return NotFound($"There is no record with the specified Id: {homeless.Id}");
 
             _repository.Update(homeless);
 
@@ -62,9 +65,13 @@ namespace HomeCompassApi.Controllers.Cases
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var homeless = _repository.GetById(id);
+
             if (homeless is null)
-                return NotFound();
+                return NotFound($"There is no record with the specified Id: {id}");
 
             _repository.Delete(id);
