+++ HomeCompassApi/Controllers/Feed/CommentController.cs
@@ -18,27 +18,42 @@ namespace HomeCompassApi.Controllers.Feed
         [HttpPost]
         public ActionResult Create(Comment comment)
         {
-            if (comment is null)
-                return BadRequest();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
 
             _repository.Add(comment);
             return CreatedAtAction(nameof(Get), new { Id = comment.Id }, comment);
         }
 
         [HttpGet]
-        public ActionResult<List<Comment>> Get() => _repository.GetAll().ToList();
+        public ActionResult<List<Comment>> Get() => Ok(_repository.GetAll().ToList());
 
 
         [HttpGet("{id}")]
-        public ActionResult<Comment> Get(int id) => _repository.GetById(id);
+        public ActionResult<Comment> Get(int id)
+        {
+            if (id <= 0)
+                return BadRequest();
+
+            var comment = _repository.GetById(id);
+
+            if (comment is null)
+                return NotFound($"There is no comment with the specified Id: {id}");
+
+            return Ok(comment);
+        }
 
 
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var comment = _repository.GetById(id);
+
             if (comment is null)
-                return NotFound();
+                return NotFound($"There is no comment with the specified Id: {id}");
 
             _repository.Delete(id);
             return NoContent();
@@ -47,8 +62,11 @@ namespace HomeCompassApi.Controllers.Feed
         [HttpPut] // ("{id}")
         public IActionResult Update(Comment comment)
         {
-            if (comment is null || _repository.GetById(comment.Id) is null)
-                return BadRequest();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            if (_repository.GetById(comment.Id) is null)
+                return NotFound($"There is no record with the specified Id: {comment.Id}");
 
             _repository.Update(comment);
             return NoContent();
